finalyearproject/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
pydantic
pydantic-settings
python-dotenv
cryptography
